packages/rag-learning-project/requirements.txt
# Core RAG dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-ollama>=0.2.0

# Vector databases
chromadb>=0.4.22
faiss-cpu>=1.7.4
pymongo>=4.6.0  # MongoDB Atlas Vector Search

# Embeddings
voyageai>=0.2.0  # Voyage AI contextualized embeddings
sentence-transformers>=2.2.2
openai>=1.10.0
langchain-text-splitters>=0.0.1  # For document chunking

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# LLM providers (choose based on your needs)
anthropic>=0.18.0
cohere>=4.47

# CLI dependencies
rich>=13.0.0
typer>=0.9.0
prompt-toolkit>=3.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=24.0.0
